[BugFix] 프로젝트 병합 과정에서 발생한 오류 해결
- API URI 충돌 해결 (FoodController에서 deleteFood와 deleteFoodFromPurchase)
- FoodController와 FoodService에서 어노테이션 및 import 오류 해결
- FoodController, FoodService에 존재하던 '공동구매' 관련 API들을 모두 PurchaseController, PurchaseService로 이동
