+++ src/main/kotlin/team/b2/bingojango/domain/food/service/FoodService.kt
@@ -1,5 +1,6 @@
 package team.b2.bingojango.domain.food.service
 
+import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import team.b2.bingojango.domain.food.model.Food
@@ -14,6 +15,8 @@ import team.b2.bingojango.domain.purchase_product.repository.PurchaseProductRepo
 import team.b2.bingojango.domain.food.dto.AddFoodRequest
 import team.b2.bingojango.domain.food.dto.UpdateFoodRequest
 import team.b2.bingojango.domain.food.model.FoodCategory
+import team.b2.bingojango.domain.food.repository.FoodRepository
+import team.b2.bingojango.domain.purchase.service.PurchaseService
 import team.b2.bingojango.domain.refrigerator.repository.RefrigeratorRepository
 import java.time.ZonedDateTime
 import team.b2.bingojango.domain.refrigerator.model.Refrigerator
@@ -29,7 +32,6 @@ import team.b2.bingojango.global.util.EntityFinder
 class FoodService(
     private val foodRepository: FoodRepository,
     private val refrigeratorRepository: RefrigeratorRepository,
-    private val purchaseService: PurchaseService,
     private val productRepository: ProductRepository,
     private val purchaseRepository: PurchaseRepository,
     private val purchaseProductRepository: PurchaseProductRepository,
@@ -76,79 +78,4 @@ class FoodService(
         if (findFood.refrigerator?.id != findRefrigerator.id){throw Exception("냉장고에 음식이 존재하지 않습니다.")}
         foodRepository.delete(findFood)
     }
-    
-    /*
-        [API] 해당 식품을 n개 만큼 공동구매 신청
-            - 검증 조건 1 : 관리자(STAFF)만 공동구매를 신청할 수 있음
-            - 검증 조건 2 : 현재 공동구매 중인 식품은 추가할 수 없음
-     */
-    fun addFoodToPurchase(userPrincipal: UserPrincipal, refrigeratorId: Long, foodId: Long, count: Int) =
-        getCurrentPurchase(userPrincipal, refrigeratorId).let {
-            if (entityFinder.getMember(userPrincipal.id, refrigeratorId).role != MemberRole.STAFF)
-                throw InvalidRoleException()
-            else if (purchaseProductRepository.findAllByPurchase(getCurrentPurchase())
-                    .map { purchaseProduct -> purchaseProduct.product.food }
-                    .contains(entityFinder.getFood(foodId))
-            ) throw AlreadyInPurchaseException()
-
-            purchaseProductRepository.save(
-                PurchaseProduct(
-                    count = count,
-                    purchase = it,
-                    product = getProduct(foodId, refrigeratorId),
-                    refrigerator = entityFinder.getRefrigerator(refrigeratorId)
-                )
-            )
-            "${entityFinder.getFood(foodId).name} ${count}개가 공동구매 신청되었습니다." // TODO: 추후 분리 예정
-        }
-
-    /*
-        [API] 공동구매 목록에서 특정 식품 삭제
-            - 검증 조건 1: 해당 공동구매를 올린 사람만 삭제를 할 수 있음
-            - 검증 조건 2: 현재 공동구매에 존재하는 식품만 삭제할 수 있음
-     */
-    fun deleteFoodFromPurchase(userPrincipal: UserPrincipal, refrigeratorId: Long, foodId: Long) {
-        if (getCurrentPurchase().proposedBy != userPrincipal.id)
-            throw InvalidRoleException()
-        purchaseProductRepository.delete(
-            purchaseProductRepository.findByRefrigeratorAndProduct(
-                refrigerator = entityFinder.getRefrigerator(refrigeratorId),
-                product = getProduct(foodId, refrigeratorId)
-            ) ?: throw ModelNotFoundException("식품")
-        )
-    }
-
-    // [내부 메서드] 현재 진행 중인(status 가 ACTIVE 한) Purchase 를 리턴 (없으면 예외 처리)
-    private fun getCurrentPurchase() =
-        purchaseRepository.findAll().firstOrNull { it.status == PurchaseStatus.ACTIVE }
-            ?: throw NoCurrentPurchaseException()
-
-    // [내부 메서드] 현재 진행 중인(status 가 ACTIVE 한) Purchase 를 리턴 (없으면 새로운 Purchase 객체 생성 후 리턴)
-    private fun getCurrentPurchase(userPrincipal: UserPrincipal, refrigeratorId: Long) =
-        purchaseRepository.findAll().firstOrNull { it.status == PurchaseStatus.ACTIVE }
-            ?: makePurchase(userPrincipal, entityFinder.getRefrigerator(refrigeratorId))
-
-    // [내부 메서드] Purchase 객체 생성
-    private fun makePurchase(userPrincipal: UserPrincipal, refrigerator: Refrigerator) =
-        purchaseRepository.save(
-            Purchase(
-                status = PurchaseStatus.ACTIVE,
-                proposedBy = userPrincipal.id,
-                refrigerator = refrigerator
-            )
-        )
-
-    // [내부 메서드] 현재 냉장고 내에 등록된 (해당 식품에 대한) Product 를 리턴 (없으면 새로운 Product 객체 생성 후 리턴)
-    private fun getProduct(foodId: Long, refrigeratorId: Long): Product =
-        productRepository.findByFoodAndRefrigerator(
-            entityFinder.getFood(foodId),
-            entityFinder.getRefrigerator(refrigeratorId)
-        )
-            ?: addProduct(entityFinder.getFood(foodId), entityFinder.getRefrigerator(refrigeratorId))
-    
-    // [내부 메서드] Product 객체 생성
-    fun addProduct(food: Food, refrigerator: Refrigerator) =
-        productRepository.save(
-            Product(food = food, refrigerator = refrigerator)
-        )
 }
