+++ src/main/kotlin/team/b2/bingojango/domain/purchase/controller/PurchaseController.kt
@@ -4,14 +4,7 @@ import io.swagger.v3.oas.annotations.Operation
 import jakarta.validation.Valid
 import org.springframework.http.ResponseEntity
 import org.springframework.security.core.annotation.AuthenticationPrincipal
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.PutMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RequestParam
-import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.bind.annotation.*
 import team.b2.bingojango.domain.purchase.service.PurchaseService
 import team.b2.bingojango.domain.vote.dto.request.VoteRequest
 import team.b2.bingojango.global.security.UserPrincipal
@@ -21,6 +14,25 @@ import team.b2.bingojango.global.security.UserPrincipal
 class PurchaseController(
     private val purchaseService: PurchaseService
 ) {
+    @Operation(summary = "특정 식품에 대한 공동구매 신청")
+    @PostMapping("/purchase/{foodId}")
+    fun addFoodToPurchase(
+        @AuthenticationPrincipal userPrincipal: UserPrincipal,
+        @PathVariable refrigeratorId: Long,
+        @PathVariable foodId: Long,
+        @RequestParam count: Int
+    ) =
+        ResponseEntity.ok().body(purchaseService.addFoodToPurchase(userPrincipal, refrigeratorId, foodId, count))
+
+    @Operation(summary = "공동구매 목록에서 특정 식품 삭제")
+    @DeleteMapping("/purchase/{foodId}")
+    fun deleteFoodFromPurchase(
+        @AuthenticationPrincipal userPrincipal: UserPrincipal,
+        @PathVariable refrigeratorId: Long,
+        @PathVariable foodId: Long
+    ) =
+        ResponseEntity.ok().body(purchaseService.deleteFoodFromPurchase(userPrincipal, refrigeratorId, foodId))
+
     @Operation(summary = "현재 공동구매 목록을 출력")
     @GetMapping
     fun showPurchase(
